Makes pidfile tracking more robust
Guard was silent when Redis started without a pidfile (due to an
unwriteable location), which means that it then had no way to further
control the process since it had no pid to work with. This change
interrogates the pidfile supplied (or defaulted to) in order to verify
that it can be written to. If not, it will display an error and opt
not to run redis-server.