Bumps version to 2.0.0
This is a major version change because the commits to make the pidfile
tracking more robust could introduce errors into working configurations
in the wild. The reason being that, though some people may have
specified pidfile locations in their Guardfile, if those locations
didn't exist or weren't writeable then redis-server would still have
been executed, happily chugging along with no way for guard-redis to
restart because the pid wasn't known.

Now, guard-redis requires that a pid be known and accessible or else
it won't start redis to begin with.